Consolidate Neo4j clients and CI gate scripts
Neo4j: Merge read-only monitoring methods (health_check, get_statistics,
execute_query) into the sbir-graph Neo4jClient, eliminating the duplicate
client in sbir_analytics/clients/. The sbir-analytics package now re-exports
from sbir-graph as the single source of truth.

CI gates: Replace 4 separate gate scripts (gate_transition_validation,
gate_transition_evidence, gate_transition_analytics, evaluate_taxonomy_checks)
with a single data-driven gate_check.py that dispatches by gate name.
