+++ .github/workflows/ci.yml
@@ -1054,17 +1054,17 @@ jobs:
         id: gate
         run: |
           set -euo pipefail
-          uv run python scripts/ci/gate_transition_validation.py
+          uv run python scripts/ci/gate_check.py transition-validation
 
       - name: Gate on evidence completeness
         run: |
           set -euo pipefail
-          uv run python scripts/ci/gate_transition_evidence.py
+          uv run python scripts/ci/gate_check.py transition-evidence
 
       - name: Gate on analytics checks
         run: |
           set -euo pipefail
-          uv run python scripts/ci/gate_transition_analytics.py
+          uv run python scripts/ci/gate_check.py transition-analytics
 
       - name: Upload validation summary
         if: always()
