+++ packages/sbir-analytics/sbir_analytics/clients/__init__.py
@@ -2,13 +2,21 @@
 
 from .dagster_client import AssetStatus, DagsterClient, RunResult
 from .metrics_collector import MetricsCollector, PipelineMetrics
-from .neo4j_client import Neo4jClient, Neo4jHealthStatus, Neo4jStatistics
+
+# Re-export Neo4j client from sbir-graph (single source of truth)
+from sbir_graph.loaders.neo4j.client import (
+    Neo4jClient,
+    Neo4jConfig,
+    Neo4jHealthStatus,
+    Neo4jStatistics,
+)
 
 __all__ = [
     "AssetStatus",
     "DagsterClient",
     "MetricsCollector",
     "Neo4jClient",
+    "Neo4jConfig",
     "Neo4jHealthStatus",
     "Neo4jStatistics",
     "PipelineMetrics",
